+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # News
 
-## v0.6.1
+## v0.6.2
 
-- Fixed a thrown error when calling `show` with the `application/pdf`, `image/png`, and `image/svg+xml` MIME types and with a `TypstString`, `TypstText`, and `Typst`
+- Fixed documents being opened during precompilation
+    - Merged [#23 by Moelf](https://github.com/jakobjpeters/Typstry.jl/pull/23)
+    - Closed [#22 by cormullion](https://github.com/jakobjpeters/Typstry.jl/issues/22)